Changed business logic

+++ app/CartingService.Domain/Commands/V1/DeleteCart/DeleteCartHandler.cs
@@ -14,6 +14,14 @@ public class DeleteCartHandler : IRequestHandler<DeleteCartCommand>
 
     public async Task Handle(DeleteCartCommand request, CancellationToken cancellationToken)
     {
+        var existingCart = await _cartsRepository.GetAsync(request.Id, cancellationToken);
+        if (existingCart == null)
+            throw new KeyNotFoundException($"Cart with id {request.Id} does not exist");
+
         await _cartsRepository.DeleteAsync(request.Id, cancellationToken);
+
+        var deletedCart = await _cartsRepository.GetAsync(request.Id, cancellationToken);
+        if(deletedCart != null)
+            throw new Exception($"Cart with id {request.Id} was not deleted");
     }
 }
